Added a Patrollig Character and Fixed Structure and names of scripts
