+++ Assets/Scripts/CameraMovement.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Camera_Movement : MonoBehaviour
+public class CameraMovement : MonoBehaviour
 {
     GameObject player;
