+++ Assets/Scripts/Character.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb5d77a33f25f1f40b04cdff60ac03a3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
