+++ Assets/Scripts/Character/CharacterController.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CharacterController : PhysicsObject
+{
+    // Add stuff here as needed asssssssssss
+
+}
