+++ Assets/Scripts/Character/CharacterController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 729ae854bc22efb4b9e615d167239e04
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
