+++ Assets/Scripts/Character/EnemyController.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyController : CharacterController
+{
+    // ASSSSSSSSSS
+
+}
