+++ Assets/Sprites/Background.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9585cab4e0a4eca41b36e07b4a9f332d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
