+++ Assets/Sprites/Characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b6c61505fb8d2654b8cc9b26e98ca4bb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
